+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  * Fixed an issue where **update-release-notes** did not generate release notes for **XDRC Templates**.
  * Fixed an issue where **upload** failed without explaining the reason.
  * Fixed an issue where **validate** did not check changed pack_metadata.json files when running using git.
+ * Added support for **xdrctemplate** to content graph.
 
 ## 1.7.8
 * Added the capability to run the MDX server in a docker container for environments without node.